==> 00-data-ingestion-and-preparation.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
import os

from utils.data_ingestion_detached_house import (
    Dict,
    load_detached_house_data,
    calc_japan_detached_house_statistics,
    extract_pref_city_code,
    extract_house_distribution,
    calc_pref_city_distribution,
    tokyo_wards,
)

ftr_pref_muni = ['pref', 'muni']

# ============================================================================ #
# Detached house data
# ============================================================================ #

file_loc = "data/ref/e008_1.xlsx"

# load fine
df = load_detached_house_data(file_loc)	

# calculate the Japanese statistics
jp_dhouse_stat = calc_japan_detached_house_statistics(df)

# caculate the distribution on prefectures and cities
df_pref_distribution, df_city_distribution = calc_pref_city_distribution(df)

# DataFrame Index of the 21 cities
city_index_tuple = df_city_distribution.set_index(['pref', 'muni']).index.to_list()

print(
    f"There are {round(jp_dhouse_stat.total / 1_000_000, 1)} Million Detached Houses in Japan"
)
print(
    f"National Share of Old Detach House: {round(jp_dhouse_stat.old_share, 2)}%",
)

# ============================================================================ #
# Load the municipal PV capacity and extract the data for 21 major cities
# ============================================================================ #

municipal_pv_cap_rpv = pd.read_csv("data/ref/municipal_pv_cap_rpv.csv")
rpv_cap_cities = municipal_pv_cap_rpv[
    municipal_pv_cap_rpv[["pref", "muni"]].apply(tuple, axis=1).isin(city_index_tuple)
]
rpv_cap_cities = rpv_cap_cities.drop(["address", "county"], axis=1).set_index(
    ["pref", "muni"]
)

rpv_cap_tokyo = municipal_pv_cap_rpv[municipal_pv_cap_rpv["muni"].isin(tokyo_wards)]
rpv_cap_tokyo = rpv_cap_tokyo.set_index("muni").drop(
    ["address", "pref", "county"], axis=1
)
rpv_cap_tokyo.index.name = None

rpv_cap_cities.loc[("東京都", "特別区部"), :] = rpv_cap_tokyo.sum()
rpv_cap_cities = rpv_cap_cities.astype(int)
rpv_cap_cities.reset_index()

# ============================================================================ #
# Load the number of rooftop PV and extract the data for 21 major cities
# ============================================================================ #

municipal_pv_no_rpv = pd.read_csv("data/ref/municipal_pv_no_rpv.csv")
rpv_no_cities = municipal_pv_no_rpv[
    municipal_pv_no_rpv[["pref", "muni"]].apply(tuple, axis=1).isin(city_index_tuple)
]
rpv_no_cities = rpv_no_cities.drop(["address", "county"], axis=1).set_index(
    ["pref", "muni"]
)

rpv_no_tokyo = municipal_pv_no_rpv[municipal_pv_no_rpv["muni"].isin(tokyo_wards)]
rpv_no_tokyo = rpv_no_tokyo.set_index("muni").drop(
    ["address", "pref", "county"], axis=1
)
rpv_no_tokyo.index.name = None

rpv_no_cities.loc[("東京都", "特別区部"), :] = rpv_no_tokyo.sum()
rpv_no_cities = rpv_no_cities.astype(int)
rpv_no_cities.reset_index()

# ============================================================================ #
# Average size of rooftop PV
# ============================================================================ #
(rpv_cap_cities/rpv_no_cities).reset_index()

# ============================================================================ #
# Municipal Demand
# ============================================================================ #

municipal_demand = pd.read_csv("data/ref/muni_demand_fy2022.csv")

demand_cities = municipal_demand[
    municipal_demand[["pref", "muni"]].apply(tuple, axis=1).isin(city_index_tuple)
]
demand_cities = demand_cities.set_index(['pref', 'muni'])
demand_tokyo = municipal_demand[municipal_demand['muni'].isin(tokyo_wards)].set_index('muni').drop('pref', axis=1)
demand_tokyo.index.name = None

demand_cities.loc[("東京都", "特別区部"), :] = demand_tokyo.sum()
demand_cities = demand_cities.astype(int)
demand_cities.reset_index()

# ============================================================================ #
# Monthly Municipal PV Generation [kWh/kw]
# ============================================================================ #

municipal_pv_output = pd.read_csv('data/ref/municipal_monthly_pv_output.csv')
municipal_pv_output = municipal_pv_output.drop('county', axis=1)
pv_out_cities = municipal_pv_output[
    municipal_pv_output[["pref", "muni"]].apply(tuple, axis=1).isin(city_index_tuple)
]
pv_out_cities = pv_out_cities.set_index(['pref', 'muni'])

pv_out_tokyo = municipal_pv_output[municipal_pv_output['muni'].isin(tokyo_wards)].set_index('muni').drop('pref', axis=1)
pv_out_tokyo.index.name = None

pv_out_cities.loc[("東京都", "特別区部"), :] = pv_out_tokyo.mean()

pv_out_cities['total'] = pv_out_cities.sum(axis=1).to_frame('pv_out')

pv_out_cities.round(2).reset_index()

# ============================================================================ #
# CONSOLIDATE DATA
# ============================================================================ #

# ============================================================================ #
# Rooftop PV Capacity
# ============================================================================ #

renamed_col = {c: f"rpv_cap_{c}" for c in rpv_cap_cities.columns}
df_clean = rpv_cap_cities.rename(columns=renamed_col).copy()

# ============================================================================ #
# No of Rooftop Installations
# ============================================================================ #

renamed_col = {c: f"rpv_no_{c}" for c in rpv_no_cities.columns}
df_temp = rpv_no_cities.rename(columns=renamed_col)
df_clean = pd.merge(df_clean, df_temp, how="right", left_index=True, right_index=True)

# ============================================================================ #
# PV Output
# ============================================================================ #

df_clean["pv_out"] = pv_out_cities["total"]

# ============================================================================ #
# Municipal Demand
# ============================================================================ #

df_clean["dmd_muni"] = demand_cities["total"]

# ============================================================================ #
# Prefectural Residential Demand
# ============================================================================ #

df_pref_energy_share = pd.read_csv("data/ref/pref_electricity_share_2021.csv")
pref_share_residential_dict = df_pref_energy_share.set_index("pref")[
    "share_residential"
].to_dict()
df_clean["pref_res_ene_share"] = df_clean.index.get_level_values("pref").map(
    pref_share_residential_dict
)

# ============================================================================ #
# City Detached House Distribution
# ============================================================================ #

df_temp = df_city_distribution.copy()
df_temp = df_temp.set_index(["pref", "muni"])
impt_col = ["old", "new"]
new_col = [f"dhouse_{ind}" for ind in impt_col]

df_temp = df_temp[impt_col]
df_temp.columns = new_col
df_clean[new_col] = df_temp[new_col]

# ============================================================================ #
# Save
# ============================================================================ #
if True:
    df_clean.reset_index().to_csv(
        "data/gen/jp_major_cities_rtpv_dataset.csv", index=False
    )

df_clean

# ============================================================================ #
# Extract hourly demand from EPCO disclosure data
# ============================================================================ #

epco_data_filename = "data/ref/japan_epco_clean_fy2022.csv"
if os.path.exists(epco_data_filename):
    epco_hourly_fy2022 = pd.read_csv(epco_data_filename, header=[0, 1], index_col=0)
    epco_hourly_fy2022.index = pd.to_datetime(epco_hourly_fy2022.index)

    epco_hourly_demand_fy2022 = epco_hourly_fy2022.xs(key="Demand", axis=1, level=1)
    epco_hourly_demand_fy2022.index.name = None
    epco_hourly_demand_fy2022.to_csv("data/ref/epco_hourly_demand_fy2022.csv")
