==> 01-city-hourly-demand.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
import os

from utils.data_ingestion_detached_house import (
    Dict,
    load_detached_house_data,
    calc_japan_detached_house_statistics,
    calc_pref_city_distribution,
    tokyo_wards,
)

from utils.japan_area_names import cities_jp_to_en, prefecture_dict_jp_to_en, epco_names
location_jp_to_en = {**cities_jp_to_en, **prefecture_dict_jp_to_en}
ftr_pref_muni = ["pref", "muni"]

# ============================================================================ #
# Detached house data
# ============================================================================ #

file_loc = "data/ref/e008_1.xlsx"

# load fine
df = load_detached_house_data(file_loc)	

# calculate the Japanese statistics
jp_dhouse_stat = calc_japan_detached_house_statistics(df)

# caculate the distribution on prefectures and cities
df_pref_distribution, df_city_distribution = calc_pref_city_distribution(df)

# DataFrame Index of the 21 cities
city_index_tuple = df_city_distribution.set_index(['pref', 'muni']).index.to_list()

# ============================================================================ #
# Municipal Demand from ENECHO (Monthly)
# ============================================================================ #

municipal_demand = pd.read_csv("data/ref/muni_demand_fy2022.csv")

demand_cities = municipal_demand[
    municipal_demand[["pref", "muni"]].apply(tuple, axis=1).isin(city_index_tuple)
]
demand_cities = demand_cities.set_index(['pref', 'muni'])
demand_tokyo = municipal_demand[municipal_demand['muni'].isin(tokyo_wards)].set_index('muni').drop('pref', axis=1)
demand_tokyo.index.name = None

demand_cities.loc[("東京都", "特別区部"), :] = demand_tokyo.sum()
demand_cities = demand_cities.astype(int)
demand_cities.rename(index=location_jp_to_en).reset_index()

# ============================================================================ #
# EPCO Jurisdictions at the Municipal Level
# ============================================================================ #

epco_muni_juris = pd.read_csv("data/ref/epco_muni_jurisdiction.csv")
epco_muni_juris_dict = epco_muni_juris.set_index(ftr_pref_muni)['epco'].to_dict()

# assign epco for each municipality
municipal_demand['epco'] = municipal_demand[ftr_pref_muni].apply(
    lambda x: epco_muni_juris_dict.get(tuple(x)), axis=1
)

# ============================================================================ #
# EPCO level demand data based on the Municipal Data from ENECHO
# ============================================================================ #

municipal_demand_by_epco = municipal_demand.drop(ftr_pref_muni, axis=1).groupby("epco").sum().div(1_000).round(
    0
).sort_values("total", ascending=False)
municipal_demand_by_epco.index.name = None
municipal_demand_by_epco

# ============================================================================ #
# Load Hourly Demand Data from EPCOs
# ============================================================================ #

epco_hourly_demand_fy2022 = pd.read_csv(
    "data/ref/epco_hourly_demand_fy2022.csv", index_col=0
)
epco_hourly_demand_fy2022.index = pd.to_datetime(epco_hourly_demand_fy2022.index)
epco_monthly_demand_fy2022 = (
    epco_hourly_demand_fy2022.resample("M")
    .sum()
    .div(1_000)
    .T.sort_values("2023-03-31", ascending=False)
)
epco_monthly_demand_fy2022.columns = municipal_demand.columns[2:14]

epco_monthly_demand_fy2022["total"] = epco_monthly_demand_fy2022.sum(axis=1)
epco_monthly_demand_fy2022

# ============================================================================ #
# Calculate the transmission loss 
# assumption: transmission loss = ratio of the reported municipal demand (user side)
# and the EPCO demand (supply side)
# ============================================================================ #

epco_tx_loss = municipal_demand_by_epco['total']/epco_monthly_demand_fy2022['total']
epco_tx_loss

# ============================================================================ #
# Apply the transmission loss 
# ============================================================================ #

epco_hourly_demand_fy2022_scaled = epco_hourly_demand_fy2022*epco_tx_loss
epco_hourly_demand_fy2022_scaled.sum(axis=0)

# ============================================================================ #
# Electricity Demand Share of each Municipality for their corresponding EPCO
# ============================================================================ #

municipal_demand_share = []
for epco, df_temp in municipal_demand.set_index(ftr_pref_muni).groupby('epco'):
	df_temp = df_temp.drop('epco',axis=1)/municipal_demand_by_epco.T[epco]/1000
	df_temp['epco'] = epco
	municipal_demand_share.append(df_temp)
municipal_demand_share = pd.concat(municipal_demand_share)
municipal_demand_share

# ============================================================================ #
# Electricity Demand Share of the Main Cities for their corresponding EPCO
# ============================================================================ #

main_city_demand_share = municipal_demand_share[municipal_demand_share.index.isin(city_index_tuple)].copy()

df_temp = municipal_demand_share.copy()
df_temp = df_temp.reset_index()
df_temp = df_temp[df_temp["muni"].isin(tokyo_wards)].drop(['pref', 'muni', 'epco'], axis=1).sum()
df_temp['epco'] = 'tokyo'

main_city_demand_share.loc[("東京都", "特別区部"), :] = df_temp.copy()
main_city_demand_share.rename(index=location_jp_to_en)

# ============================================================================ #
# Load the Prefectural Energy Electricity Share
# ============================================================================ #

df_pref_elec_share = pd.read_csv("data/ref/pref_electricity_share_2021.csv")
pref_elec_resi_share_dict = df_pref_elec_share.set_index('pref')['share_residential'].to_dict()
df_pref_elec_share.head(5)

# ============================================================================ #
# Calculate the Residential Share relative to the EPCO demand
# ============================================================================ #

main_city_epco_elec_resi_share = main_city_demand_share.copy()
main_city_epco_elec_resi_share = main_city_epco_elec_resi_share.reset_index().set_index(
    ftr_pref_muni + ["epco"]
)
for index, row in main_city_epco_elec_resi_share.iterrows():
	main_city_epco_elec_resi_share.loc[index] = main_city_epco_elec_resi_share.loc[index]*pref_elec_resi_share_dict[index[0]]
main_city_epco_elec_resi_share.rename(index=location_jp_to_en)

# ============================================================================ #
# Calculate the residential energy demand of each main cities
# ============================================================================ #

main_city_resi_elec_dmd = []
for index, row in main_city_epco_elec_resi_share.iterrows():
    epco_name = index[2]
    muni_name = index[1]
    muni_resi_share = row["total"]
    main_city_resi_elec_dmd.append(
        (epco_hourly_demand_fy2022_scaled[epco_name] * muni_resi_share).to_frame(muni_name)
    )
main_city_resi_elec_dmd = pd.concat(main_city_resi_elec_dmd, axis=1)
main_city_resi_elec_dmd.to_csv("data/gen/major_city_resi_elec_dmd.csv")
main_city_resi_elec_dmd.rename(columns=location_jp_to_en)
